+++ src/data/preprocessing/preprocess.py
@@ -10,19 +10,8 @@ class Preprocess(metaclass = ABCMeta):
     
     def __init__(self):
         self.text_column = 'text'
-        self.file_name = None
         self.data = None
         
-    @property
-    @abstractmethod
-    def file_name(self):
-        pass
- 
-    @file_name.setter
-    @abstractmethod
-    def file_name(self,value):
-        pass
-
     @property
     @abstractmethod
     def data(self):