+++ src/data/preprocessing/preprocess_bert_doc_emb.py
@@ -4,6 +4,7 @@
 # In[3]:
 
 #src module
+from src.enums import Feature
 from src.data.preprocessing.preprocess import Preprocess
 from src import utilities as u
 from src import utilities_preprocessing as pre
@@ -13,17 +14,7 @@ class PreprocessBertDocEmb(Preprocess):
     '''Preporcesses data for BERT document embedding features.'''
     def __init__(self):
         super().__init__()
-        self.file_name = 'preprocessed_data_bert_doc_emb.csv'
-        self.data = None
 
-    @property
-    def file_name(self):
-        return self._file_name
- 
-    @file_name.setter
-    def file_name(self,value):
-        self._file_name = value
-        
     @property
     def data(self):
         return self._data 
@@ -36,4 +27,4 @@ class PreprocessBertDocEmb(Preprocess):
         #TODO
         print('\nSTARTED: Preprocessing bert document embeddings started.\n')
         
-        return self.data, self.file_name
+        return self.data