+++ src/factory.py
@@ -1,0 +1,9 @@
+
+# coding: utf-8
+
+# In[1]:
+
+def get_object(objects, name: object = None) -> object:
+    '''Factory'''
+    return objects[name]()
+