+++ src/features/build_sentiment_features.py
@@ -13,6 +13,7 @@ class BuildSentimentFeature(BuildFeature):
     '''Extracts sentiment features by using VADER sentiment analysis model.'''
     def __init__(self):
         self.sid = SentimentIntensityAnalyzer()
+        self.score_names=['neg', 'neu', 'pos', 'compound']
         
     def get_polarity_scores(self, text):
         '''Retrieves combineed polarity scores of text (neutral and compoound scores). 
@@ -24,13 +25,14 @@ class BuildSentimentFeature(BuildFeature):
         dictionary: Combined polarity scores (neutral and compoound).
     
         Example:
-            >>> get_polarity_scores("The file is super cool.")
+            >>> bsf = BuildSentimentFeature(score_names=['neu', 'compound'])
+            >>> bsf.get_polarity_scores("The file is super cool.")
             [0.326, 0.7351]
         '''
         
         #Example scores = {'neg': 0.0, 'neu': 0.326, 'pos': 0.674, 'compound': 0.7351}
         scores = self.sid.polarity_scores(text)
-        return [scores['neu'], scores['compound']]  #TODO : use self._score_names
+        return [scores[name] for name in self.score_names]
     
     def fit(self, x, y=None):
         return self